feat: add SEO component and integrate into various layouts and pages
- Introduced a new Seo component for managing document metadata and structured data.
- Integrated Seo component into NotFound, Authenticated, Brand, and various route components.
- Removed the SearchMetadata component as its functionality is now covered by the new Seo component.
- Updated multiple documentation pages to include SEO metadata for better indexing and search visibility.
- Enhanced the AdminOperations page with a utility function for formatting operation timestamps.
