+++ sdk/tests/cli/test_main.py
@@ -1,4 +1,5 @@
 import runpy
+import pytest
 from click.testing import CliRunner
 from longlink.cli.main import main
 
@@ -17,7 +18,7 @@ def test_cli_help_lists_all_supported_commands() -> None:
     assert "migrate" in result.output
 
 
-def test_module_execution_invokes_cli_entrypoint(monkeypatch) -> None:
+def test_module_execution_invokes_cli_entrypoint(monkeypatch: pytest.MonkeyPatch) -> None:
     """Run the CLI when the SDK package is executed as a module."""
 
     # Arrange
