+++ web/src/components/breadcrumb/Page.tsx
@@ -8,18 +8,19 @@ export function PageBreadcrumb({ applicationName }: { applicationName?: string }
     const { pathname } = useLocation();
     const organization = pathname.split('/')[2] ?? '';
     const label = pathname.startsWith('/admin/') ? 'Admin' : startCase(decodePathSegment(organization));
+    const isApplicationBreadcrumb = applicationName !== undefined;
     return (
         <Breadcrumbs separator=">" variant="supporting">
             <BreadcrumbItem href="/user/organizations">
                 <Wordmark />
             </BreadcrumbItem>
             <BreadcrumbItem
-                href={applicationName !== undefined ? `/orgs/${organization}` : undefined}
-                isCurrent={applicationName === undefined}
+                href={isApplicationBreadcrumb ? `/orgs/${organization}` : undefined}
+                isCurrent={!isApplicationBreadcrumb}
             >
                 {label}
             </BreadcrumbItem>
-            {applicationName !== undefined && <BreadcrumbItem isCurrent>{applicationName}</BreadcrumbItem>}
+            {isApplicationBreadcrumb && <BreadcrumbItem isCurrent>{applicationName}</BreadcrumbItem>}
         </Breadcrumbs>
     );
 }
