+++ web/src/components/layouts/Article.tsx
@@ -1,3 +1,4 @@
+import { Seo } from '@/components/Seo';
 import { dateFormatter } from '@/lib/utils';
 import { Link } from '@astryxdesign/core/Link';
 import { Text } from '@astryxdesign/core/Text';
@@ -17,6 +18,10 @@ import { Layout, LayoutContent, LayoutHeader } from '@astryxdesign/core/Layout';
 
 type ArticlePage = {
     lastUpdated: string;
+    seo: {
+        description: string;
+        title: string;
+    };
     toc?: Array<{ id: string; label: string; level: number }>;
     editUrl?: string;
 };
@@ -90,84 +95,93 @@ export function Article({ children, page }: { children: ReactNode; page: Article
     }, [currentPage]);
 
     return (
-        <Layout
-            height="auto"
-            header={
-                <LayoutHeader className="sticky top-14 z-20 bg-card lg:top-2" hasDivider padding={0}>
-                    <Stack className="relative" height={64} width="100%">
-                        <PageContainer height="100%" justify="center" maxWidth={1064} paddingInline={6}>
-                            <Breadcrumb className="min-w-0 overflow-hidden" />
-                        </PageContainer>
-                        <Center className="absolute end-0 top-0" height={64} paddingInline={4}>
-                            <Button href="/user/organizations" label="Get Started" size="sm" variant="primary" />
-                        </Center>
-                    </Stack>
-                </LayoutHeader>
-            }
-            content={
-                <LayoutContent isScrollable={false} padding={6}>
-                    <Stack className="mx-auto" direction="horizontal" gap={6} maxWidth={1016} width="100%">
-                        <PageContainer className="min-w-0" maxWidth={768}>
-                            <article className="article-content space-y-7 text-justify">
-                                {children}
-                                <Stack as="footer" gap={3}>
-                                    {currentPage >= 0 ? (
-                                        <Stack
-                                            aria-label="Documentation page navigation"
-                                            direction="horizontal"
-                                            hAlign="between"
-                                            width="100%"
-                                        >
-                                            <Button
-                                                aria-keyshortcuts="ArrowLeft"
-                                                href={previousPage}
-                                                icon={<ArrowLeft aria-hidden size={16} />}
-                                                isDisabled={previousPage === undefined}
-                                                label="Previous"
-                                                onClick={scrollToArticleTop}
-                                            />
-                                            <Button
-                                                aria-keyshortcuts="ArrowRight"
-                                                endContent={<ArrowRight aria-hidden size={16} />}
-                                                href={nextPage}
-                                                isDisabled={nextPage === undefined}
-                                                label="Next"
-                                                onClick={scrollToArticleTop}
-                                            />
-                                        </Stack>
-                                    ) : null}
-                                    <Divider />
-                                    <Stack direction="horizontal" gap={3} hAlign="between" vAlign="center" wrap="wrap">
-                                        <Text type="supporting">
-                                            {`Last updated: ${dateFormatter.format(new Date(page.lastUpdated))}`}
-                                        </Text>
-                                        {page.editUrl ? (
-                                            <Link href={page.editUrl} hasUnderline isExternalLink type="supporting">
-                                                Edit this page in GitHub
-                                            </Link>
+        <>
+            <Seo {...page.seo} hasBreadcrumbs />
+            <Layout
+                height="auto"
+                header={
+                    <LayoutHeader className="sticky top-14 z-20 bg-card lg:top-2" hasDivider padding={0}>
+                        <Stack className="relative" height={64} width="100%">
+                            <PageContainer height="100%" justify="center" maxWidth={1064} paddingInline={6}>
+                                <Breadcrumb className="min-w-0 overflow-hidden" />
+                            </PageContainer>
+                            <Center className="absolute end-0 top-0" height={64} paddingInline={4}>
+                                <Button href="/user/organizations" label="Get Started" size="sm" variant="primary" />
+                            </Center>
+                        </Stack>
+                    </LayoutHeader>
+                }
+                content={
+                    <LayoutContent isScrollable={false} padding={6}>
+                        <Stack className="mx-auto" direction="horizontal" gap={6} maxWidth={1016} width="100%">
+                            <PageContainer className="min-w-0" maxWidth={768}>
+                                <article className="article-content space-y-7 text-justify">
+                                    {children}
+                                    <Stack as="footer" gap={3}>
+                                        {currentPage >= 0 ? (
+                                            <Stack
+                                                aria-label="Documentation page navigation"
+                                                direction="horizontal"
+                                                hAlign="between"
+                                                width="100%"
+                                            >
+                                                <Button
+                                                    aria-keyshortcuts="ArrowLeft"
+                                                    href={previousPage}
+                                                    icon={<ArrowLeft aria-hidden size={16} />}
+                                                    isDisabled={previousPage === undefined}
+                                                    label="Previous"
+                                                    onClick={scrollToArticleTop}
+                                                />
+                                                <Button
+                                                    aria-keyshortcuts="ArrowRight"
+                                                    endContent={<ArrowRight aria-hidden size={16} />}
+                                                    href={nextPage}
+                                                    isDisabled={nextPage === undefined}
+                                                    label="Next"
+                                                    onClick={scrollToArticleTop}
+                                                />
+                                            </Stack>
                                         ) : null}
+                                        <Divider />
+                                        <Stack
+                                            direction="horizontal"
+                                            gap={3}
+                                            hAlign="between"
+                                            vAlign="center"
+                                            wrap="wrap"
+                                        >
+                                            <Text type="supporting">
+                                                {`Last updated: ${dateFormatter.format(new Date(page.lastUpdated))}`}
+                                            </Text>
+                                            {page.editUrl ? (
+                                                <Link href={page.editUrl} hasUnderline isExternalLink type="supporting">
+                                                    Edit this page in GitHub
+                                                </Link>
+                                            ) : null}
+                                        </Stack>
                                     </Stack>
+                                </article>
+                            </PageContainer>
+                            {page.toc?.length ? (
+                                <Stack
+                                    as="aside"
+                                    aria-label="On this page"
+                                    className="sticky top-20 hidden shrink-0 self-start lg:flex"
+                                    gap={3}
+                                    padding={5}
+                                    width={224}
+                                >
+                                    <Text type="label" weight="semibold">
+                                        On this page
+                                    </Text>
+                                    <Outline items={page.toc} density="compact" label="On this page" />
                                 </Stack>
-                            </article>
-                        </PageContainer>
-                        {page.toc?.length ? (
-                            <Stack
-                                as="aside"
-                                aria-label="On this page"
-                                className="sticky top-20 hidden shrink-0 self-start lg:flex"
-                                gap={3}
-                                padding={5}
-                                width={224}
-                            >
-                                <Text type="label" weight="semibold">
-                                    On this page
-                                </Text>
-                                <Outline items={page.toc} density="compact" label="On this page" />
-                            </Stack>
-                        ) : null}
-                    </Stack>
-                </LayoutContent>
-            }
-        />
+                            ) : null}
+                        </Stack>
+                    </LayoutContent>
+                }
+            />
+        </>
     );
 }
