+++ web/src/platform/root.tsx
@@ -1,7 +1,6 @@
 import { Outlet } from 'react-router';
 import { RootProvider } from '@/providers';
 import '@/index.css';
-import { SearchMetadata } from '@/components/layouts/SearchMetadata';
 
 export { Document as Layout } from '@/components/layouts/Document';
 
@@ -9,7 +8,6 @@ export { Document as Layout } from '@/components/layouts/Document';
 export default function PlatformRoot() {
     return (
         <RootProvider>
-            <SearchMetadata />
             <Outlet />
         </RootProvider>
     );
