+++ web/src/platform/routes/docs/api/Applications.tsx
@@ -9,6 +9,10 @@ import { Article } from '@/components/layouts/Article';
 const applicationPaths = ['Use', 'Adapt', 'Create'];
 
 const metadata = {
+    seo: {
+        title: 'Applications | Platform Documentation | LongLink',
+        description: 'Learn how to create, deploy, and operate applications on the LongLink Platform.',
+    },
     toc: [{ id: 'applications', label: 'Applications', level: 1 }],
     lastUpdated: '2026-08-05',
     editUrl: 'https://github.com/xLongLink/longlink/edit/main/web/src/platform/routes/docs/api/Applications.tsx',
